+++ src/main/java/org/openlmis/referencedata/repository/custom/FacilityRepositoryCustom.java
@@ -16,10 +16,11 @@
 package org.openlmis.referencedata.repository.custom;
 
 import org.openlmis.referencedata.domain.Facility;
+import org.openlmis.referencedata.domain.FacilityType;
 import org.openlmis.referencedata.domain.GeographicZone;
 
 import java.util.List;
 
 public interface FacilityRepositoryCustom {
-  List<Facility> search(String code, String name, GeographicZone zone);
+  List<Facility> search(String code, String name, GeographicZone zone, FacilityType facilityType);
 }